Optimize fetching build status
When a build succeeds, we can get enough information from the list of packages
that we don't need to query again for the full build status.

I've confirmed that this produces the same json output as before.  The only
downside is we need to manually construct the result url, so if COPR changes
the url format this code will need to be updated.

For COPR projects with ~1500 packages, just patch reduced the time to
generate the json from 20 minutes to 5 minutes.